+++ src/ResourceManager/RecoveryServices/Commands.RecoveryServices.Backup.Models/CmdletParamEnums.cs
@@ -50,11 +50,24 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.Models
     public enum RestoreBackupItemParams
     {
         RecoveryPoint,
-        StorageAccountId,
-        StorageAccountLocation,
-        StorageAccountType,
+        StorageAccountName,
+        StorageAccountResourceGroupName
+    }
+
+    public enum RestoreVMBackupItemParams
+    {
         TargetResourceGroupName,
         OsaOption
+    }
+
+    public enum RestoreFSBackupItemParams
+    {
+        ResolveConflict,
+        SourceFilePath,
+        SourceFileType,
+        TargetStorageAccountName,
+        TargetFileShareName,
+        TargetFolder
     }
 
     public enum PolicyParams
@@ -71,7 +84,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.Models
 
     public enum ItemParams
     {
-        AzureVMName,
+        ItemName,
         AzureVMCloudServiceName,
         AzureVMResourceGroupName,
         WorkloadType,
@@ -84,6 +97,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.Models
         DeleteBackupData,
         BackupManagementType,
         ExpiryDateTimeUTC,
+        StorageAccountName,
     }
 
     public enum ProtectionCheckParams